+++ sei-ibc-go/modules/apps/transfer/keeper/msg_server_test.go
@@ -1,0 +1,71 @@
+package keeper_test
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	"github.com/cosmos/ibc-go/v3/modules/apps/transfer/types"
+)
+
+func (suite *KeeperTestSuite) TestMsgTransfer() {
+	var msg *types.MsgTransfer
+
+	testCases := []struct {
+		name     string
+		malleate func()
+		expPass  bool
+	}{
+		{
+			"success",
+			func() {},
+			true,
+		},
+		{
+			"invalid sender",
+			func() {
+				msg.Sender = "address"
+			},
+			false,
+		},
+		{
+			"sender is a blocked address",
+			func() {
+				msg.Sender = suite.chainA.GetSimApp().AccountKeeper.GetModuleAddress(types.ModuleName).String()
+			},
+			false,
+		},
+		{
+			"channel does not exist",
+			func() {
+				msg.SourceChannel = "channel-100"
+			},
+			false,
+		},
+	}
+
+	for _, tc := range testCases {
+		suite.SetupTest()
+
+		path := NewTransferPath(suite.chainA, suite.chainB)
+		suite.coordinator.Setup(path)
+
+		coin := sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(100))
+		msg = types.NewMsgTransfer(
+			path.EndpointA.ChannelConfig.PortID,
+			path.EndpointA.ChannelID,
+			coin, suite.chainA.SenderAccount.GetAddress().String(), suite.chainB.SenderAccount.GetAddress().String(),
+			suite.chainB.GetTimeoutHeight(), 0, // only use timeout height
+		)
+
+		tc.malleate()
+
+		res, err := suite.chainA.GetSimApp().TransferKeeper.Transfer(sdk.WrapSDKContext(suite.chainA.GetContext()), msg)
+
+		if tc.expPass {
+			suite.Require().NoError(err)
+			suite.Require().NotNil(res)
+		} else {
+			suite.Require().Error(err)
+			suite.Require().Nil(res)
+		}
+	}
+}